+++ cookbooks/yum/attributes/remi.rb
@@ -1,8 +1,7 @@
 #
 # Cookbook Name:: yum
-# Attributes:: default
+# Attributes:: remi
 #
-# Copyright 2011, Eric G. Wolfe
 # Copyright 2011, Opscode, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,13 +17,14 @@
 # limitations under the License.
 #
 
-# Example: override.yum.exclude = "kernel* compat-glibc*"
-default['yum']['exclude'] = Array.new
-default['yum']['installonlypkgs'] = Array.new
-default['yum']['ius_release'] = '1.0-11'
-default['yum']['repoforge_release'] = '0.5.2-2'
-default['yum']['proxy'] = ''
-default['yum']['proxy_username'] = ''
-default['yum']['proxy_password'] = ''
-default['yum']['cachedir'] = '/var/cache/yum'
-default['yum']['keepcache'] = 0
+case node['platform']
+when "fedora"
+  default['yum']['remi']['url'] = "http://rpms.famillecollet.com/fedora/#{node['platform_version'].to_i}/remi/mirror"
+else
+  default['yum']['remi']['url'] = "http://rpms.famillecollet.com/enterprise/#{node['platform_version'].to_i}/remi/mirror"
+end
+
+default['yum']['remi']['key'] = "RPM-GPG-KEY-remi"
+default['yum']['remi']['key_url'] = "http://rpms.famillecollet.com/#{node['yum']['remi']['key']}"
+default['yum']['remi']['includepkgs'] = nil
+default['yum']['remi']['exclude'] = nil