+++ cookbooks/yum/metadata.rb
@@ -1,8 +1,9 @@
+name              "yum"
 maintainer        "Opscode, Inc."
 maintainer_email  "cookbooks@opscode.com"
 license           "Apache 2.0"
 long_description  IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version           "1.0.0"
+version           "2.2.4"
 recipe            "yum", "Empty recipe."
 recipe            "yum::yum", "Manages yum configuration"