+++ cookbooks/yum/recipes/elrepo.rb
@@ -1,9 +1,8 @@
 #
 # Cookbook Name:: yum
-# Attributes:: default
+# Recipe:: elrepo
 #
-# Copyright 2011, Eric G. Wolfe
-# Copyright 2011, Opscode, Inc.
+# Copyright:: Copyright (c) 2013 Opscode, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,15 +15,17 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
 
-# Example: override.yum.exclude = "kernel* compat-glibc*"
-default['yum']['exclude'] = Array.new
-default['yum']['installonlypkgs'] = Array.new
-default['yum']['ius_release'] = '1.0-11'
-default['yum']['repoforge_release'] = '0.5.2-2'
-default['yum']['proxy'] = ''
-default['yum']['proxy_username'] = ''
-default['yum']['proxy_password'] = ''
-default['yum']['cachedir'] = '/var/cache/yum'
-default['yum']['keepcache'] = 0
+yum_key node['yum']['elrepo']['key'] do
+  url  node['yum']['elrepo']['key_url']
+  action :add
+end
+
+yum_repository "elrepo" do
+  description "ELRepo.org Community Enterprise Linux Extras Repository"
+  key node['yum']['elrepo']['key']
+  mirrorlist node['yum']['elrepo']['url']
+  includepkgs node['yum']['elrepo']['includepkgs']
+  exclude node['yum']['elrepo']['exclude']
+  action :create
+end