+++ cookbooks/yum/CHANGELOG.md
@@ -1,3 +1,59 @@
+## v2.2.4:
+
+### Bug
+
+- [COOK-2360]: last commit to `yum_repository` changes previous behaviour
+- [COOK-3015]: Yum cookbook test minitest to fail
+
+## v2.2.2:
+
+### Improvement
+
+- [COOK-2741]: yum::elrepo
+- [COOK-2946]: update tests, test kitchen support in yum cookbook
+
+### Bug
+
+- [COOK-2639]: Yum cookbook - epel - always assumes url is a mirror list
+- [COOK-2663]: Yum should allow metadata_expire setting in repo file
+- [COOK-2751]: Update yum.ius_release version to 1.0-11
+
+## v2.2.0:
+
+* [COOK-2189] - yum::ius failed on install (caused from rpm dependency)
+* [COOK-2196] - Make includepkgs and exclude configurable for each repos
+* [COOK-2244] - Allow configuring caching using attributes
+* [COOK-2399] - yum cookbook LWRPs fail FoodCritic
+* [COOK-2519] - Add priority option to Yum repo files
+* [COOK-2593] - allow integer or string for yum priority
+* [COOK-2643] - don't use conditional attribute for `yum_key` `remote_file`
+
+## v2.1.0:
+
+* [COOK-2045] - add remi repository recipe
+* [COOK-2121] - add `:create` action to `yum_repository`
+
+## v2.0.6:
+
+* [COOK-2037] - minor style fixes
+* [COOK-2038] - updated README
+
+## v2.0.4:
+
+* [COOK-1908] - unable to install repoforge on CentOS 6 32 bit
+
+## v2.0.2:
+
+* [COOK-1758] - Add default action for repository resource
+
+## v2.0.0:
+
+This version changes the behavior of the EPEL recipe (most commonly
+used in other Opscode cookbooks) on Amazon, and removes an attribute,
+`node['yum']['epel_release']`. See the README for details.
+
+* [COOK-1772] - Simplify management of EPEL with LWRP
+
 ## v1.0.0:
 
 `mirrorlist` in the `yum_repository` LWRP must be set to the mirror